WIP: backend implementation of switch variable resolution, tests
