+++ packages/dara-core/dara/core/main.py
@@ -314,6 +314,8 @@ def _start_application(config: Configuration):
                     template_registry.register(name, renderer(config))
 
     except Exception as e:
+        import traceback
+        traceback.print_exc()
         dev_logger.error(
             'Something went wrong when building application template, there is most likely an issue in the application logic',
             e,
